Made a base RUDAT_13G_90 class. To create a USB driver, simply mixin the MiniCircuitsHIDMixin
