+++ qcodes/instrument_drivers/Minicircuits/USBHIDMixin.py
@@ -3,6 +3,7 @@ A mixin module for USB Human Interface Device instruments
 """
 import os
 import time
+import struct
 
 try:
     import pywinusb.hid as hid
@@ -43,7 +44,6 @@ class USBHIDMixin(Instrument):
     def __init__(self, name, instance_id: str=None, timeout: float=2,
                  **kwargs) ->None:
 
-        super().__init__(name, **kwargs)
         self._check_hid_import()
 
         devs = hid.HidDeviceFilter(
@@ -63,6 +63,8 @@ class USBHIDMixin(Instrument):
         self._data_buffer: bytes = None
         self._timeout = timeout
         self._device.set_raw_data_handler(self._handler)
+
+        super().__init__(name, **kwargs)
 
     def _handler(self, data: bytes) ->None:
         self._data_buffer = data
@@ -132,3 +134,63 @@ class USBHIDMixin(Instrument):
         ).get_devices()
 
         return [dev.instance_id for dev in devs]
+
+
+class MiniCircuitsHIDMixin(USBHIDMixin):
+    """
+    The specific implementation for mini circuit human interface devices
+
+    Args:
+        name (str)
+        instance_id (str): The id of the instrument we want to connect. If
+            there is only one instrument then this is an optional argument.
+            If we have more then one instrument, quarry their ID's by calling
+            the class method 'enumerate_devices'
+        timeout (float): Specify a timeout for this instrument
+    """
+    vendor_id = 0x20ce
+    product_id = 0x0023
+
+    def __init__(self, name: str, instance_id: str=None, timeout: float=2,
+                 **kwargs) ->None:
+
+        self._usb_endpoint = 0
+        self._end_of_message = b"\x00"
+        self.packet_size = 64
+
+        super().__init__(name, instance_id, timeout, **kwargs)
+
+    def _pack_string(self, cmd: str) ->bytes:
+        """
+        Pack a string to a binary format such that it can be send to the
+        HID.
+
+        Args:
+            cmd (str)
+        """
+        str_len = len(cmd)
+        pad_len = self.packet_size - str_len
+
+        if pad_len < 0:
+            raise ValueError(f"Length of data exceeds {self.packet_size} B")
+
+        command_number = 1
+        packed_data = struct.pack(
+            f"BB{str_len}s{pad_len}x",
+            self._usb_endpoint, command_number, cmd.encode("ascii")
+        )
+
+        return packed_data
+
+    def _unpack_string(self, response: bytes) ->str:
+        """
+        Unpack data from the instrument to a string
+
+        Args:
+            response (bytes)
+        """
+        _, _, reply_data = struct.unpack(
+            f"BB{self.packet_size-1}s", bytes(response)
+        )
+        span = reply_data.find(self._end_of_message)
+        return reply_data[:span].decode("ascii")
